project/requirements.txt
asgiref==3.3.1
asprise-ocr-sdk-python-api==15.3.0.8
backcall==0.2.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-environ==0.4.5
django-extensions==3.0.9
django-extra-fields==3.0.2
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.3.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
Pygments==2.7.2
pytesseract==0.3.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
tinyocr==0.2.3
traitlets==5.0.5
wcwidth==0.2.5
whitenoise==5.2.0
